unify post new + edit into PostForm component
